split noit module into 2 pieces to create a simple reversable connection module
