+++ .gitignore
@@ -94,3 +94,4 @@ test/test-stratcon.csr
 test/test-stratcon.key
 test/t/logs
 pkg/debian/changelog
+*~
