--- src/js/index.js
@@ -1,7 +1,0 @@
-exports.log = require('./log');
-exports.connection_set_stats = require('./connection').set_stats;
-exports.connection = require('./connection').connection;
-exports.setCA = require('./connection').setCA;
-exports.firehose = require('./firehose');
-exports.mapper = require('./map_uuid').mapper;
-exports.utils = require('./utils');
