+++ src/js/noit-connection/.npmignore
@@ -1,0 +1,1 @@
+*~
