+++ src/js/noit-connection/index.js
@@ -1,0 +1,4 @@
+exports.connection_set_stats = require('./noit-connection').set_stats;
+exports.connection = require('./noit-connection').noit_connection;
+exports.setCA = require('./noit-connection').setCA;
+exports.utils = require('./utils');
