+++ src/js/noit/histogram.js
