+++ src/js/noit/package.json
@@ -1,6 +1,6 @@
 {
   "name": "noit",
-  "version": "0.0.4",
+  "version": "0.0.5",
   "description": "Noit Connection Utilities",
   "main": "index.js",
   "repository": {
@@ -13,6 +13,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "noit-connection": "^0.0.1",
     "protocol-buffers": "^3.1.6",
     "fq": "^0.3.1",
     "pg": "^4.1.1",
